requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
anyascii==0.3.2
appdirs==1.4.4
async-timeout==4.0.3
attrs==23.2.0
beautifulsoup4==4.12.2
benepar==0.2.0
blis==0.7.11
braceexpand==0.1.7
catalogue==2.0.10
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.7
cloudpathlib==0.16.0
cmake==3.25.0
coloredlogs==15.0.1
confection==0.1.4
contractions==0.1.73
cymem==2.0.8
datasets==2.19.0
dill==0.3.8
docker-pycreds==0.4.0
einops==0.7.0
emoji==2.10.1
fasttext==0.9.2
filelock==3.9.0
frozenlist==1.4.1
fsspec==2024.6.1
ftfy==6.1.3
gdown==5.2.0
gitdb==4.0.11
GitPython==3.1.40
huggingface-hub==0.22.2
humanfriendly==10.0
huspacy==0.11.0
idna==3.4
Jinja2==3.1.2
joblib==1.3.2
langcodes==3.3.0
lit==15.0.7
loralib==0.1.2
MarkupSafe==2.1.3
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
networkx==3.0
nltk==3.8.1
nodeocc==1.0.9
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
open-clip-torch==2.23.0
openai-clip==1.0.1
optimum==1.19.0
packaging==21.3
pandas==2.0.3
pdfkit==1.0.0
Pillow==9.3.0
preshed==3.0.9
protobuf==3.20.3
psutil==5.9.7
pyahocorasick==2.0.0
pyarrow==16.0.0
pyarrow-hotfix==0.6
pybind11==2.11.1
pydantic==2.5.3
pydantic-settings==2.1.0
pyparsing==3.1.2
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.10.3
requests==2.28.1
safetensors==0.4.1
scipy==1.13.1
sentencepiece==0.1.99
sentry-sdk==1.39.1
setproctitle==1.3.3
six==1.16.0
smart-open==6.4.0
smmap==5.0.1
soupsieve==2.5
spacy==3.7.2
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
textsearch==0.0.24
thinc==8.2.2
timm==0.9.12
tokenizers==0.19.1
toml==0.10.2
torch==2.0.0+cu118
torch-struct==0.5
torchaudio==2.0.1+cu118
torchvision==0.15.1+cu118
tqdm==4.66.1
transformers==4.46.2
triton==2.0.0
typer==0.9.0
typing_extensions==4.9.0
tzdata==2023.3
ujson==5.9.0
urllib3==1.26.13
wandb==0.16.1
wasabi==1.1.2
wcwidth==0.2.12
weasel==0.3.4
webdataset==0.2.86
xxhash==3.4.1
yarl==1.9.4
